+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Add numParams() method to Network
     - Neural Network Learners now report number of trainable parameters
     - Regex Filter added pattern to match unicode emojis
+    - Custom escape character for CSV Extractor
 
 - 2.3.3
     - Optimize Adam and AdaMax Optimizers